mainにDevelopmentをマージ

+++ db/migrate/20231019081331_create_algos.rb
@@ -1,7 +1,7 @@
 class CreateAlgos < ActiveRecord::Migration[7.0]
   def change
     create_table :algos do |t|
-      t.bigint :algo_id
+      t.integer :algo_id
       t.string :algo_name
 
       t.timestamps
